requirements.txt
fastapi
uvicorn
pika
python-multipart
watchdog
google-generativeai
pytesseract
pdf2image
Pillow
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
python-dotenv
sqlalchemy
websockets